[modules-core] Extract common Kotlin utilities
