+++ packages/expo-image-picker/android/src/main/java/expo/modules/imagepicker/ImagePickerOptions.kt
@@ -1,6 +1,7 @@
 package expo.modules.imagepicker
 
 import expo.modules.core.Promise
+import expo.modules.core.utilities.ifNull
 
 data class ImagePickerOptions(
   val quality: Int,
