+++ packages/expo-media-library/android/src/main/java/expo/modules/medialibrary/assets/CreateAsset.kt
@@ -10,12 +10,12 @@ import android.net.Uri
 import android.os.Build
 import androidx.annotation.RequiresApi
 import expo.modules.core.Promise
+import expo.modules.core.utilities.ifNull
 import expo.modules.medialibrary.ERROR_IO_EXCEPTION
 import expo.modules.medialibrary.ERROR_NO_FILE_EXTENSION
 import expo.modules.medialibrary.ERROR_UNABLE_TO_LOAD_PERMISSION
 import expo.modules.medialibrary.ERROR_UNABLE_TO_SAVE
 import expo.modules.medialibrary.MediaLibraryUtils
-import expo.modules.medialibrary.ifNull
 import java.io.File
 import java.io.FileInputStream
 import java.io.FileOutputStream
