+++ packages/expo-modules-core/android/src/main/java/expo/modules/core/utilities/KotlinUtilities.kt
@@ -1,4 +1,4 @@
-package expo.modules.medialibrary
+package expo.modules.core.utilities
 
 /**
  * Returns receiver, or block result if the receiver is `null`
@@ -16,8 +16,8 @@ inline fun <T> T?.ifNull(block: () -> T): T = this ?: block()
  *
  * Works the same as the `as?` operator, but allows method chaining without parentheses:
  * ```
- *   val x = a.b?.takeIfInstanceOf<Number>?.someMethod()
- *   val y = (a.b? as? Number)?.someMethod() // same, but needs parenthesis
+ *   val x = a.b.takeIfInstanceOf<Number>?.someMethod()
+ *   val y = (a.b as? Number)?.someMethod() // same, but needs parenthesis
  * ```
  */
 inline fun <reified T> Any?.takeIfInstanceOf(): T? = if (this is T) this else null
